requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.2
django-audit-fields==0.3.9
django-crypto-fields @ git+https://github.com/botswana-harvard/django-crypto-fields.git@f975c27f7ea69ca274976a3868c34f2d8983d550
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-ipware==5.0.0
django-js-reverse==0.9.1
django-logentry-admin==1.1.0
django-picklefield==3.1
django-q==1.3.9
django-revision @ git+https://github.com/botswana-harvard/django-revision.git@38e3c9f8c07de4977fdd8f73166562253a9ac281
django-simple-history==3.3.0
docutils==0.19
edc-action-item @ git+https://github.com/botswana-harvard/edc-action-item.git@022e2a98a76a8c0f506733e8e196667331e77a5b
edc-appointment @ git+https://github.com/botswana-harvard/edc-appointment.git@5c8a64f53b54db7a057a296c769d4e69f3400a3b
edc-base @ git+https://github.com/botswana-harvard/edc-base.git@08a64f9aa1a17809f8c3bcc6e1ccdb247cdd9880
edc-calendar @ git+https://github.com/botswana-harvard/edc-calendar.git@612d892114d769db17b5e030def5d44e11221edc
edc-consent @ git+https://github.com/botswana-harvard/edc-consent.git@90b71fd80bc9191268d32215b0d20230eb845ddd
edc-constants @ git+https://github.com/botswana-harvard/edc-constants.git@70b03cc3c135f242de543123b3c7c9dfce284e67
edc-dashboard @ git+https://github.com/botswana-harvard/edc-dashboard.git@fbef227eaa7951c58e2d97937741fca370210f70
edc-device @ git+https://github.com/botswana-harvard/edc-device.git@867751966a4449ffd17719aad11bba6be2a52800
edc-facility @ git+https://github.com/botswana-harvard/edc-facility.git@3cd82c74b28ec22d16c4d627eff1aaf311af4287
edc-fieldsets @ git+https://github.com/botswana-harvard/edc-fieldsets.git@a11157ca78f69fe5c074f18f3073a849e0d2c82d
edc-form-validators @ git+https://github.com/botswana-harvard/edc-form-validators.git@17e6eaf4cbef1b80ec76721f905a9d76a66d9fe9
edc-identifier @ git+https://github.com/botswana-harvard/edc-identifier.git@c849cdb598d87082ab2008d7855bd1710e7023de
edc-lab @ git+https://github.com/botswana-harvard/edc-lab.git@6cfecbc2032d00b4cb4cfd27a2550d3d0dd7524d
edc-label @ git+https://github.com/botswana-harvard/edc-label.git@8fe927071ac87507f5d85603cef6c4bbd8a508f9
edc-locator @ git+https://github.com/botswana-harvard/edc-locator.git@607caa33dbd1ae445a45bc387a0ce1421bddd414
edc-metadata @ git+https://github.com/botswana-harvard/edc-metadata.git@cc0cf2e47d7ba5d074423a49c910682e33503318
edc-metadata-rules @ git+https://github.com/botswana-harvard/edc-metadata-rules.git@aaf50edb81fa3d201b471d93e154ecdd1cf5dd69
edc-model-admin @ git+https://github.com/botswana-harvard/edc-model-admin.git@18da70d85b1bf921fa50f48d432bb4b8ab2523b4
edc-model-wrapper @ git+https://github.com/botswana-harvard/edc-model-wrapper.git@814b9f85feceed246e378425ab6add8bba3b655f
edc-prn @ git+https://github.com/botswana-harvard/edc-prn.git@ff7d32cdfafc4ee225602931860c38c268148ae3
edc-protocol @ git+https://github.com/botswana-harvard/edc-protocol.git@eb9bce9c4e832543c9171d1fabacd25b5db25f90
edc-reference @ git+https://github.com/botswana-harvard/edc-reference.git@cf77d5877476342830366e91a3fe6cfd4758945e
edc-registration @ git+https://github.com/botswana-harvard/edc-registration.git@3c44c67996e21d1c79c41b54befb44d6deba67ac
edc-search @ git+https://github.com/botswana-harvard/edc-search.git@9d60b5b606b903245fad0004e432279df52da776
edc-senaite-interface @ git+https://github.com/botswana-harvard/edc-senaite-interface.git@83fa26d800386227ea2263611bfec2aae4803e32
edc-sms @ git+https://github.com/botswana-harvard/edc-sms.git@488e609263bbccbeaa85da7a3adf49132304d323
edc-subject-dashboard @ git+https://github.com/botswana-harvard/edc-subject-dashboard.git@2933ac9ce87f7a9dd5698f5e6cead12d0205aad8
edc-timepoint @ git+https://github.com/botswana-harvard/edc-timepoint.git@2231996b42606a422ef4763c3ab388cddf38abbe
edc-utils==0.3.19
edc-visit-schedule @ git+https://github.com/botswana-harvard/edc-visit-schedule.git@71588bc1e9f1e5218ab524940b11c4ec406fee9f
edc-visit-tracking @ git+https://github.com/botswana-harvard/edc-visit-tracking.git@61f7ccd1ce3f4b078a389b844029ef1263469ec2
Faker==18.5.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
install==1.3.5
model-mommy==2.0.0
pycparser==2.21
pycryptodome==3.17
pycups==2.0.1
PyMySQL==1.0.3
python-dateutil==2.8.2
python-memcached==1.59
pytz==2023.3
redis==3.5.3
requests==2.29.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
toolz==0.12.0
tqdm==4.65.0
urllib3==1.26.15
wcwidth==0.2.6
xlwt==1.3.0
